Azure/codes/monitoring/requirements.txt
pandas
evidently
joblib
azure-storage-blob
prometheus-client
pyarrow
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation